create auth and file data access
